At this point stereo is working. in iris/src/DSO, I copied immersiveWindow.cpp too    hdImmersiveWindow.cpp     - a generic HD monitor    fourkImmersiveWindow.cpp  - a generic 4K monitor    arcImmersiveWindow.cpp    - the NCAT Arc6 display
Only hdImmersive is complete. It generated 6 side (DSO) like the original
immersiveWindow did for the CAVE. The problem is its basic assumption that
edges of the screen touch, for the Arc, I want the core cube to be centered
one the front wall. The Arc only has a front. I could not get it centered as I
wanted. I experimented with using the old Boulder Powewall DSO as the model
for these three display types and can be centered.

After this commit, I will implement based on the powerwall DSO. Note, this
assumed the display is flat The Arc6 is not flat, so this is only a short term
solution.  I later expect to come back to multi-DSO approach, but instead of
90 degrees between walls, I'll be using the Arc6 angle and locaton of its 6
"stacks".